requirements-test.txt
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
pytest-cov==4.1.0
logfire==2.8.0
logfire-api==2.8.0
pydantic-ai==0.0.14
aiohttp==3.9.3
